% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_estimates}
\alias{county_estimates}
\title{county_estimates}
\format{
A data frame with 5481 rows and 5 variables:
\describe{
\item{\code{Lan}}{character Swedish county}
\item{\code{Datum}}{double Date}
\item{\code{Uppskattning}}{double Prediction of covid for county based on reported symptoms}
\item{\code{Low_CI}}{double Confidence Interval - lower bound}
\item{\code{High_CI}}{double Confidence Interval - higher bound}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
county_estimates
}
\description{
COVID Symptom Study data - county level
}
\keyword{datasets}
