% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidsymptom.R
\docType{package}
\name{covidsymptom}
\alias{covidsymptom}
\alias{covidsymptom-package}
\title{covidsymptom: COVID Symptom Study-Sweden data}
\description{
Easily imports COVID Symptom Study-Sweden open data
}
\author{
\strong{Maintainer}: Hugo Fitipaldi \email{hugo.fitipaldi@med.lu.se}

}
