% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{csss_tests}
\alias{csss_tests}
\title{Daily covid-19 tests reported in the COVID Symptom Study Sweden}
\format{
A data frame of 4 columns
\describe{
\item{date}{Date of report}
\item{total_tests}{Number of tests}
\item{proportion_positive}{Proportion of positive tests}
\item{test_type}{Type of test: snabbtest (antigen) or PCR}
}
}
\source{
COVID Symptom Study-Sweden.
}
\usage{
csss_tests
}
\description{
A dataset containing daily reported number of covid-19 tests and the proportion of positive results
}
\details{
This dataset contains daily reports of covid-19 tests and the proportion of positive results
in swedish 2-digit postcode regions.
}
\examples{
data(csss_tests)
}
\keyword{COVID19}
\keyword{coronavirus}
\keyword{datasets}
