% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_data.R
\name{get_latest_data}
\alias{get_latest_data}
\title{Retrieve updated COVID Symptom Study Sweden (CSSS) data}
\usage{
get_latest_data(data_level = c("national", "county", "postcode"))
}
\arguments{
\item{data_level}{selects which data set from CSSS you want to download: "national" (default), "county" or "postcode"}
}
\value{
This function returns the selected \code{data.frame}
}
\description{
Downloads latest version of CSSS data.
}
\details{
This function was designed to download the latest data from COVID Symptom Study Sweden without the need to update the package on a regular basis.
This function has been disabled and will no longer download data, as the COVID Symptom Study Sweden
is no longer updating their data sets. Please refrain from using this function.
}
\examples{
df <- get_latest_data(data_level = "county")

}
\author{
Hugo Fitipaldi
}
