% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{national_estimates}
\alias{national_estimates}
\title{national_estimates}
\format{
A data frame with 261 rows and 4 variables:
\describe{
\item{\code{Datum}}{double Date}
\item{\code{Uppskattning}}{double Prediction (\%) of covid for Sweden based on reported symptoms}
\item{\code{Low_CI}}{double Confidence Interval - lower bound}
\item{\code{High_CI}}{double Confidence Interval - higher bound}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
national_estimates
}
\description{
COVID Symptom Study data - national level
}
\keyword{datasets}
