% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{postcode_estimates}
\alias{postcode_estimates}
\title{postcode_estimates}
\format{
A data frame with 22427 rows and 5 variables:
\describe{
\item{\code{Postnummer}}{integer 2-digit postal code region}
\item{\code{Datum}}{double Date}
\item{\code{Uppskattning}}{double Prediction of covid for 2-digit region based on reported symptoms}
\item{\code{Low_CI}}{double Confidence Interval - lower bound}
\item{\code{High_CI}}{double Confidence Interval - higher bound}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
postcode_estimates
}
\description{
COVID Symptom Study data - 2-digit postcode level
}
\keyword{datasets}
