% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{symptoms}
\alias{symptoms}
\title{Daily proportion of symptoms reported by CSSS users}
\format{
A data frame of 4 columns
\describe{
\item{Datum}{Date of report}
\item{Symptom}{Symptom}
\item{Andel}{Proportion of symptom}
\item{Vikt}{Weight related to covid19: positiv or negativ}
}
}
\source{
COVID Symptom Study-Sweden.
}
\usage{
symptoms
}
\description{
A dataset containing daily reported number of covid-19 tests and the proportion of positive results
}
\details{
This dataset contains daily reports of covid-19 tests and the proportion of positive results
in swedish 2-digit postcode regions.
}
\examples{
data(csss_tests)
}
\keyword{COVID19}
\keyword{coronavirus}
\keyword{datasets}
