% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_estimates}
\alias{county_estimates}
\title{Estimates of COVID-19 in Sweden (county level)}
\format{
A data frame of 5 columns
\describe{
\item{Lan}{Swedish county}
\item{Datum}{Date of prediction}
\item{Uppskattning}{Estimated prevalence of COVID-19 in Sweden}
\item{Low_CI}{Confidence Interval Lower Bound}
\item{High_CI}{Confidence Interval Higher Bound}
}
}
\source{
COVID Symptom Study-Sweden.
}
\usage{
county_estimates
}
\description{
A dataset containing predictions of COVID Symptom Study Sweden at national level
}
\details{
This dataset contains daily estimated prevalence of symptomatic covid-19
in swedish counties.
}
\examples{
data(county_estimates)
}
\keyword{COVID19}
\keyword{coronavirus}
\keyword{datasets}
