% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidsymptom.R
\docType{package}
\name{covidsymptom}
\alias{covidsymptom}
\alias{covidsymptom-package}
\title{covidsymptom: COVID Symptom Study-Sweden data}
\description{
Easily imports COVID Symptom Study-Sweden open data
}
\details{
The COVID Symptom Study is a non-commercial project that uses a free smartphone app to
facilitate real-time data collection of symptoms, exposures, and risk factors related to COVID-19.
The app was developed by researchers at King’s College and Guys and St Thomas’ Hospitals in London in
partnership with health science company Zoe Global Ltd. Baseline data and recurring daily questions are
described in Drew et al (Science, 2020). The app was launched in the UK and US March 2020. In Sweden,
the study is based at Lund University and, as per a collaboration agreement on 28 July 2020, Uppsala
University. The app was launched in Sweden on April 29, 2020 as part of a national research initiative
on COVID-19. To date, >4.5 million participants in the three countries are using the app, ~202,000 of
whom live in Sweden. Participants have so far made ~271 million data entries, with Swedish participants
contributing ~11 million of these.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/csss-resultat/covidsymptom}
}

}
\author{
\strong{Maintainer}: Hugo Fitipaldi \email{hugo.fitipaldi@med.lu.se}

}
