% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_data.R
\name{get_latest_data}
\alias{get_latest_data}
\title{Function to retrieve latest COVID Symptom Study - Sweden datasets}
\usage{
get_latest_data(data_level = c("national", "county", "postcode"))
}
\arguments{
\item{data_level}{selects which dataset from CSSS you want to download}
}
\value{
selected dataset
}
\description{
This function imports the later CSSS datasets
}
