% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{national_estimates}
\alias{national_estimates}
\title{Estimates of COVID-19 in Sweden (national level)}
\format{
A data frame of 4 columns
\describe{
\item{Datum}{Date of prediction}
\item{Uppskattning}{Estimated prevalence of COVID-19 in Sweden}
\item{Low_CI}{Confidence Interval Lower Bound}
\item{High_CI}{Confidence Interval Higher Bound}
}
}
\source{
COVID Symptom Study-Sweden
}
\usage{
national_estimates
}
\description{
A dataset containing predictions of COVID Symptom Study Sweden at national level
}
\keyword{datasets}
