% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{postcode_estimates}
\alias{postcode_estimates}
\title{Estimates of COVID-19 in Sweden (2-digit postal code level)}
\format{
A data frame of 5 columns
\describe{
\item{Postnummer}{2-digit postal code area}
\item{Datum}{Date of prediction}
\item{Uppskattning}{Estimated prevalence of COVID-19 in Sweden}
\item{Low_CI}{Confidence Interval Lower Bound}
\item{High_CI}{Confidence Interval Higher Bound}
\item{Ort}{Regions/cities within the postcode area }
}
}
\source{
COVID Symptom Study-Sweden \href{https://www.covid19app.lu.se/}{website}.
}
\usage{
postcode_estimates
}
\description{
A dataset containing predictions of COVID Symptom Study Sweden at national level
}
\details{
This dataset contains daily estimated prevalence of symptomatic covid-19
in swedish 2-digit postcode regions.
}
\examples{
data(postcode_estimates)
}
\keyword{COVID19}
\keyword{coronavirus}
\keyword{datasets}
